backend/requirements.txt
fastapi==0.128.2
uvicorn==0.40.0
pymongo==4.16.0
python-dotenv==1.2.1
pydantic[email]==2.12.5
email-validator==2.3.0
